Allow selecting the audio driver
Fuse has many different audio drivers but the one that is enabled is
selected automatically by the configure script depending on the
operating system and/or the UI library.

In Linux there are several drivers (alsa, oss, libao, sdl) and the
user has only partial control about which one to enable with options
like --without-alsa or --with-libao.

This patch gives the user complete control by replacing these options
with a new one (--with-audio-driver=DRV), and the list of available
drivers is printed by the configure script. If this option is unset a
default driver is selected using the same criteria as before.

This also allows the user to select the SDL audio driver independently
from the SDL UI.