Fix detection of some audio drivers
Here's how to tell which one of these drivers to enable:

1) Solaris has <sys/audio.h> and defines AUDIO_SETINFO
2) HP/UX has <sys/audio.h> and defines AUDIO_FORMAT_LINEAR16BIT
3) OpenBSD does not have <sys/audio.h> but has <sys/audioio.h>

However the configure check is broken for the first two because it
looks for dsound.h instead, which is from Microsoft's DirectSound.

Another problem is that those drivers are in completely different
files (sunsound.c and hpsound.c) but both are built and we're using
ifdefs to check for the aforementioned macros and decide which
implementation gets compiled.

This is unnecessary because the configure script can already decide
what file needs to be built, and it is also wrong because the macros
are checked before even including the headers that define them.

It seems that all this was messed up when the sound selection logic
was moved from sound/lowlevel.h to configure.in (patch #133).

This patch fixes these problems so the correct drivers can be built.
Whether they are all still working in modern versions of those
operating systems is a different question, at least OpenBSD removed
API used by the Fuse driver two years ago, see audioio.h r1.26 in
https://github.com/openbsd/src/commit/9215aa3dfad387bca877a805534df6dc