Use AM_PATH_SDL only once in configure.ac
Fuse can use the SDL library in two different cases that can be
configured independently:

1) User interface + audio driver
2) Joystick support

The configure.ac script uses AM_PATH_SDL in both cases in order to
detect if SDL is available.

This is not a big problem in itself, but if we want to decouple the UI
and audio driver options (to allow for example GTK+ UI with SDL audio,
or SDL UI with ALSA audio) we would need to add AM_PATH_SDL to the
audio driver configuration as well, complicating the logic a bit more.

This patch makes configure.ac use AM_PATH_SDL only once.