Allow using SDL 2 for the audio and joystick drivers
The Fuse SDL UI only supports SDL 1 at the moment, but the audio and
joystick drivers can work with SDL 2 just fine.

After this patch Fuse selects SDL 2 by default if available unless the
SDL UI is selected or the configure script is run with --disable-sdl2.

The joystick driver needs one small change: SDL_INIT_VIDEO should not
be used with non-SDL UIs. This seems to be harmless with SDL 1, but
crashes the program with SDL 2.

(patch #430)